font: give the auto-hinter a CFF glyph's points exactly as FreeType reads them
The hinter has to see the points FreeType's loader stores, or it hints a
different glyph. For CFF fonts with fractional coordinates (David CLM,
Frank Ruehl CLM) it did not, in three ways:

* Precision. Charstring operands are 16.16 fixed point -- 32 bits, where
  f32 keeps 24 -- and FreeType decides where a coordinate lands, and
  whether a contour closes, in 1024ths of a unit. The interpreter now does
  its arithmetic in f64 (exact for 16.16), and the hinter's points are a
  new exact type, sfnt::Exact; only the rasterizer's path narrows to f32.
* Which points exist. FreeType's cf2 engine drops a line of no length in
  device space (1024ths of a unit, not just exact zero), stores a
  contour's first point only once something is drawn from it, and folds a
  contour's last point into its first where they meet in device space.
  sfnt::CffPoints records a glyph under those rules; the interpreter draws
  on a CffPen -- the Outline path or CffPoints -- and Cff::tagged_outline
  gives the hinter the latter. seac accents are drawn from their offset,
  as FreeType draws them, since where points land decides which lines
  vanish; a nested seac is refused, as FreeType refuses it.
* Units. FreeType floors a CFF coordinate to whole units after snapping it
  to 1024ths (Units::Floored, measured on 17,788 points: all agree), and
  scales x from those whole units with FT_MulFix. The hinter now does both,
  so x is exact too, for TrueType as well.

Every glyph of both CFF fonts now agrees with FreeType, both coordinates
exact; TrueType fonts agree as before, now with x exact rather than within
a 64th. The golden fixture records x as well as y, and gains `w`, drawn in
16.16 fractions to exercise flooring, both kinds of vanishing line and
both kinds of near-closure (mutation-checked: rounding instead of flooring,
or keeping short lines, each fails 18 rows). hint_oracle.py compares x
exactly, breaks differences down into points / y / x and by style, and no
longer overflows a Windows command line on a large font.

+++ gui/font/src/scaled.rs
@@ -424,7 +424,7 @@ impl ScaledFont {
     /// **For diagnostics only.** Drawing never needs it; it is the half of the
     /// check against FreeType's auto-hinter (`tools/hint_oracle.py`) that this
     /// crate answers, and FreeType answers in stored points, not paths.
-    pub fn hinted_points(&mut self, gid: u16) -> Option<Vec<(f32, f32, bool)>> {
+    pub fn hinted_points(&mut self, gid: u16) -> Option<Vec<(f64, f64, bool)>> {
         self.hinter()?;
         let Hinting::Ready(Some(hinter)) = &self.hinting else {
             return None;
